requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-phonenumber-field==5.2.0
gunicorn==20.1.0
phonenumbers==8.12.27
phonenumberslite==8.12.27
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0
